requirements.txt
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
pillow==10.0.0
tqdm==4.66.1
flask==2.3.3
werkzeug==2.3.7
requests==2.31.0
